Move response struct under v2
Also removedd `/metrics/*namespace` endpoint, because we obtain the same
feature with `/metrics` and the parameter "ns" in the query.
